requirements.txt
python-dotenv
Flask
discord.py
pyyaml
requests
openai
lyricsgenius